+++ app/commands/HelpCommand.php
@@ -1,7 +1,6 @@
 <?php
 
-use System\Apps\Command;
-
+use System\Console\Command;
 use System\Apps\CLI;
 
 class HelpCommand extends Command