+++ bootstrap/init.php
@@ -12,7 +12,6 @@ require_once dirname(__DIR__) . '/app/core/RouterProvider.php';
 require_once dirname(__DIR__) . '/app/core/Controller.php';
 require_once dirname(__DIR__) . '/app/core/CLI.php';
 require_once dirname(__DIR__) . '/app/core/Service.php';
-require_once dirname(__DIR__) . '/app/core/Command.php';
 require_once dirname(__DIR__) . '/app/core/GlobalFuntion.php';
 
 // Declare Config Class