+++ tests/BasicTest.php
@@ -17,7 +17,6 @@ final class BasicTest extends TestCase
     $this->assertFileExists('./bootstrap/autoload.php');
 
     $this->assertFileExists('./app/core/CLI.php');
-    $this->assertFileExists('./app/core/Command.php');
     $this->assertFileExists('./app/core/Config.php');
     $this->assertFileExists('./app/core/Controller.php');
     $this->assertFileExists('./app/core/GlobalFuntion.php');
@@ -26,6 +25,7 @@ final class BasicTest extends TestCase
     $this->assertFileExists('./app/core/RouterProvider.php');
     $this->assertFileExists('./app/core/Service.php');
 
+    $this->assertFileExists('./app/core/template/command');
     $this->assertFileExists('./app/core/template/controller');
     $this->assertFileExists('./app/core/template/model');
     $this->assertFileExists('./app/core/template/service');